Add audit trail for high priority system changes
Committing a change to running, copying to a datastore, or calling an
RPC may be restricted by NACM.  In these cases a system administrator
may want to know who made a change or sent an RPC.

For commit, we are only interested in changes to the running ds, which
is when the system actually activates the changes.  Copying to startup
or other datastores is handled separately.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>